pyproject.toml
[project]
name = "adk-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=0.2.0",
    "litellm>=1.67.0.post1",
    "pillow>=11.2.1",
]
